+++ CHANGES.rst
@@ -1,3 +1,9 @@
+0.0.2
+-----
+
+* When not provided, ``authenticated_userid`` and ``effective_principals`` from
+  super class ``CallbackAuthenticationPolicy`` are used.
+
 0.0.1
 -----